+++ altrepo_api/api/site_packageset/sql.py
@@ -157,6 +157,90 @@ GROUP BY pkg_name
 ORDER BY pkg_name
 """
 
+    get_find_packages_by_name_and_arch = """
+WITH
+lp_preselect AS
+(
+    SELECT
+        pkg_hash,
+        pkgset_name
+    FROM static_last_packages
+    WHERE pkg_name ILIKE '%{name}%'
+        AND pkg_sourcepackage = 1
+        {branch}
+),
+srcs_by_arch AS
+(
+    SELECT
+        pkg_hash,
+        pkgset_name
+    FROM lp_preselect
+    WHERE pkg_hash IN (
+        SELECT pkg_srcrpm_hash
+        FROM Packages
+        WHERE pkg_sourcepackage = 0
+            AND pkg_srcrpm_hash IN (
+                SELECT pkg_hash FROm lp_preselect
+            )
+            {arch}
+    )
+),
+lp_preselect2 AS
+(
+    SELECT
+        pkg_hash,
+        pkgset_name
+    FROM static_last_packages
+    WHERE pkg_name NOT ILIKE '%{name}%'
+        AND pkg_sourcepackage = 1
+        {branch}
+)
+SELECT
+    pkg_name,
+    groupUniqArray((LP.pkgset_name, pkg_version, pkg_release, toString(pkg_hash))),
+    max(pkg_buildtime),
+    argMax(pkg_url, pkg_buildtime),
+    argMax(pkg_summary, pkg_buildtime),
+    any(pkg_group_)
+FROM Packages
+INNER JOIN srcs_by_arch AS LP USING (pkg_hash)
+WHERE pkg_hash IN
+(
+    SELECT pkg_hash FROM srcs_by_arch
+)
+GROUP BY pkg_name
+ORDER BY pkg_name
+UNION ALL
+SELECT
+    pkg_name,
+    groupUniqArray((LP2.pkgset_name, pkg_version, pkg_release, toString(pkg_hash))),
+    max(pkg_buildtime),
+    argMax(pkg_url, pkg_buildtime),
+    argMax(pkg_summary, pkg_buildtime),
+    any(pkg_group_)
+FROM Packages
+INNER JOIN lp_preselect2 AS LP2 USING (pkg_hash)
+WHERE pkg_name NOT ILIKE '%{name}%'
+    AND pkg_sourcepackage = 1
+    AND pkg_sourcerpm IN
+    (
+        SELECT pkg_sourcerpm
+        FROM Packages
+        WHERE pkg_sourcepackage = 0
+            AND pkg_name ILIKE '%{name}%'
+            {arch}
+            AND pkg_hash IN (
+                SELECT pkg_hash
+                FROM static_last_packages
+                WHERE pkg_sourcepackage = 0
+                    {branch}
+            )
+    )
+    {branch}
+GROUP BY pkg_name
+ORDER BY pkg_name
+"""
+
     get_find_deleted_packages_by_name = """
 WITH
 deleted_src_pkgs AS (
@@ -355,6 +439,37 @@ WHERE pkgset_uuid IN (
         AND pkg_version = '{version}'
         AND pkg_release = '{release}'
 )
+"""
+
+    get_source_pkg_name = """
+SELECT DISTINCT
+    pkg_name
+FROM Packages
+WHERE pkg_hash IN (
+    SELECT pkg_srcrpm_hash
+    FROM Packages
+    WHERE pkg_hash IN (
+        SELECT pkg_hash
+        FROM static_last_packages
+        WHERE pkg_name = '{name}'
+            AND pkgset_name = '{branch}'
+        )
+)
+"""
+
+    get_src_pkg_by_bin = """
+SELECT DISTINCT
+    pkg_name
+FROM BranchPackageHistory
+WHERE pkg_hash IN (
+    SELECT pkg_srcrpm_hash
+    FROM Packages
+    WHERE pkg_name = '{name}'
+        AND pkg_sourcepackage = 0
+    ORDER BY pkg_srcrpm_hash ASC
+) AND pkg_sourcepackage = 1
+  AND pkgset_name = '{branch}'
+ORDER BY task_changed DESC LIMIT 1
 """