+++ altrepo_api/api/task_progress/endpoints/last_tasks.py
@@ -20,10 +20,19 @@ from altrepo_api.api.base import APIWorker
 from ..sql import sql
 
 
-@dataclass(frozen=True, order=True)
-class IterationMeta:
-    task_try: int
-    task_iter: int
+@dataclass
+class TaskApprovalMeta:
+    task_id: int
+    date: datetime.datetime
+    type: str
+    nickname: str
+    message: str
+
+
+@dataclass
+class SubtaskArchsMeta:
+    arch: str
+    stage_status: str
 
 
 @dataclass
@@ -42,10 +51,7 @@ class SubtaskMeta:
     subtask_pkg_from: str
     subtask_changed: datetime.datetime
     type: str
-    stage: str
-    stage_status: str
-    status: str
-    archs: list[str] = field(default_factory=list)
+    archs: list[SubtaskArchsMeta] = field(default_factory=list)
 
 
 @dataclass
@@ -54,13 +60,14 @@ class TaskMeta:
     task_repo: str
     task_state: str
     task_owner: str
+    task_try: int
+    task_iter: int
     task_changed: str
     task_message: str
-    iterations: list[IterationMeta] = field(default_factory=list)
+    task_stage: str
+    dependencies: list[int] = field(default_factory=list)
     subtasks: list[SubtaskMeta] = field(default_factory=list)
-
-    def __post_init__(self):
-        self.iterations = sorted([IterationMeta(*el) for el in self.iterations], reverse=True)
+    approval: list[TaskApprovalMeta] = field(default_factory=list)
 
 
 class LastTasks(APIWorker):
@@ -121,11 +128,78 @@ class LastTasks(APIWorker):
 
         tasks = [TaskMeta(*el) for el in response]  # type: ignore
 
-        # get subtask info by task_id
+        # get task approval info by task_id
+        _tmp_table = "tmp_task_ids"
         response = self.send_sql_request(
-            self.sql.get_subtasks.format(
-                tasks=[el.task_id for el in tasks],
-            )
+            self.sql.get_task_approval.format(tmp_table=_tmp_table),
+            external_tables=[
+                {
+                    "name": _tmp_table,
+                    "structure": [
+                        ("task_id", "UInt32"),
+                    ],
+                    "data": [
+                        {
+                            "task_id": el.task_id,
+                        }
+                        for el in tasks
+                    ],
+                }
+            ],
+        )
+        if not self.sql_status:
+            return self.error
+        if response:
+            approval = [TaskApprovalMeta(*el[:-1]) for el in response]
+            for task in tasks:
+                task.approval = [el for el in approval if el.task_id == task.task_id]
+
+        # get task depends if task_state == "POSTPONED"
+        _tmp_table = "tmp_task_ids_postponed"
+        response = self.send_sql_request(
+            self.sql.get_task_dependencies.format(tmp_table=_tmp_table),
+            external_tables=[
+                {
+                    "name": _tmp_table,
+                    "structure": [
+                        ("task_id", "UInt32"),
+                    ],
+                    "data": [
+                        {
+                            "task_id": el.task_id,
+                        }
+                        for el in tasks
+                        if el.task_state == "POSTPONED"
+                    ],
+                }
+            ],
+        )
+        if not self.sql_status:
+            return self.error
+        if response:
+            for task in tasks:
+                for el in response:
+                    if el[0] == task.task_id:
+                        task.dependencies = el[1]
+
+        # get subtask info by task_id
+        _tmp_table = "tmp_task_ids"
+        response = self.send_sql_request(
+            self.sql.get_subtasks.format(tmp_table=_tmp_table),
+            external_tables=[
+                {
+                    "name": _tmp_table,
+                    "structure": [
+                        ("task_id", "UInt32"),
+                    ],
+                    "data": [
+                        {
+                            "task_id": el.task_id,
+                        }
+                        for el in tasks
+                    ],
+                }
+            ],
         )
         if not self.sql_status:
             return self.error
@@ -136,29 +210,42 @@ class LastTasks(APIWorker):
                     "args": self.args,
                 }
             )
-
         subtasks = [SubtaskMeta(*el) for el in response]
 
         # get subtask archs by task_id and subtask_id
+        tasks_subtasks = []
+        for task in tasks:
+            for sub in subtasks:
+                if (
+                    task.task_state in ("BUILDING", "FAILED", "FAILING")
+                    and task.task_id == sub.task_id
+                ):
+                    tasks_subtasks.append(
+                        {"task_id": sub.task_id, "subtask_id": sub.subtask_id}
+                    )
+
+        _tmp_table = "tmp_tasks_subtasks"
         response = self.send_sql_request(
-            self.sql.get_subtasks_archs.format(
-                subtasks=[(el.task_id, el.subtask_id) for el in subtasks],
-            )
+            self.sql.get_subtasks_status.format(tmp_table=_tmp_table),
+            external_tables=[
+                {
+                    "name": _tmp_table,
+                    "structure": [
+                        ("task_id", "UInt32"),
+                        ("subtask_id", "UInt32"),
+                    ],
+                    "data": tasks_subtasks,
+                }
+            ],
         )
         if not self.sql_status:
             return self.error
-        if not response:
-            return self.store_error(
-                {
-                    "message": "No data found in database",
-                    "args": self.args,
-                }
-            )
-
-        for subtask in subtasks:
-            for el in response:
-                if el[0] == subtask.task_id and el[1] == subtask.subtask_id:
-                    subtask.archs = el[2]
+        if response:
+            for subtask in subtasks:
+                for el in response:
+                    if el[0] == subtask.task_id and el[1] == subtask.subtask_id:
+                        subtask.archs = [SubtaskArchsMeta(*arch) for arch in el[2]]
+                        subtask.type = el[3]
 
         for task in tasks:
             task.subtasks = [el for el in subtasks if el.task_id == task.task_id]