+++ altrepo_api/api/task_progress/serializers.py
@@ -20,6 +20,13 @@ from .namespace import get_namespace
 
 ns = get_namespace()
 
+subtask_archs_model = ns.model(
+    "SubTaskArchitecturesModel",
+    {
+        "stage_status": fields.String(description="stage status"),
+        "arch": fields.String(description="subtask architecture"),
+    },
+)
 subtasks_el_model = ns.model(
     "SubTasksElementModel",
     {
@@ -38,17 +45,19 @@ subtasks_el_model = ns.model(
         "subtask_pkg_from": fields.String(description="subtask package from"),
         "subtask_changed": fields.DateTime(description="subtask changed"),
         "type": fields.String(description="subtask type"),
-        "stage": fields.String(description="subtask stage"),
-        "stage_status": fields.String(description="stage status"),
-        "status": fields.String(description="subtask status"),
-        "archs": fields.List(fields.String(description="subtask architecture list")),
+        "archs": fields.Nested(
+            subtask_archs_model, description="list of subtask architectures"
+        ),
     },
 )
-task_iterations_el_model = ns.model(
-    "TaskIterationsElementModel",
+task_approval_el_model = ns.model(
+    "TaskApprovalElementModel",
     {
-        "task_try": fields.Integer(description="task try number"),
-        "task_iter": fields.Integer(description="task iteration number"),
+        "task_id": fields.Integer(description="task id"),
+        "date": fields.DateTime(description="approval date"),
+        "type": fields.String(description="approval type"),
+        "nickname": fields.String(description="maintainer nickname"),
+        "message": fields.String(description="approval message"),
     },
 )
 last_tasks_el_model = ns.model(
@@ -58,13 +67,19 @@ last_tasks_el_model = ns.model(
         "task_repo": fields.String(description="repository name"),
         "task_state": fields.String(description="task state"),
         "task_owner": fields.String(description="task owner"),
+        "task_try": fields.Integer(description="task try number"),
+        "task_iter": fields.Integer(description="task iteration number"),
         "task_changed": fields.DateTime(description="task changed"),
         "task_message": fields.String(description="task message"),
-        "iterations": fields.Nested(
-            task_iterations_el_model, description="task iteration list", as_list=True
-        ),
+        "task_stage": fields.String(description="task stage"),
+        "dependencies": fields.List(fields.Integer, description="task dependencies"),
         "subtasks": fields.Nested(
             subtasks_el_model, description="list of subtasks by task", as_list=True
+        ),
+        "approval": fields.Nested(
+            task_approval_el_model,
+            description="list of approvals for task",
+            as_list=True,
         ),
     },
 )
@@ -78,5 +93,13 @@ last_tasks_model = ns.model(
             description="list of latest task changes",
             as_list=True,
         ),
+    },
+)
+
+all_pkgsets_model = ns.model(
+    "AllPackageSetsModel",
+    {
+        "length": fields.Integer(description="number of packagesets found"),
+        "branches": fields.List(fields.String, description="list of packagesets"),
     },
 )